requirements.txt
streamlit
pandas
langchain
langchain-community
langchain-google-genai
pypdf
docx2txt
faiss-cpu
sentence-transformers
requests
python-dotenv
duckduckgo-search
langgraph
groq
langchain-groq
langchain-openai
langchain-huggingface
rank_bm25
